requirements.txt
fastapi
pydantic
uvicorn
elasticsearch
sentence-transformers
langchain_community
langchain-openai
langchain-core
openai
dotenv